DUOS-1554 added tooltip for legacy RP vote, modularized component into smaller sub-components for import
